Clean up, threads and commands
Moved visualization into SDL thread and made simple cli parser.
Threads can terminate each other by SDL_SemPost(running).
