+++ model/robot.cpp
@@ -1,0 +1,1 @@
+#include "robot.h"
