+++ model/robot.h
@@ -1,0 +1,9 @@
+#ifndef __MODEL_ROBOT_H__
+#define __MODEL_ROBOT_H__
+
+class Robot {
+	
+};
+
+/* __MODEL_ROBOT_H__ */
+#endif
